+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 ### Added
+- `shrink_to_fit()` method on `World` for shrinking the `World`'s current allocation to the minimum required for the current data.
 - `ContainsComponent`, `ContainsEntities`, `ContainsEntity`, `ContainsParQuery`, and `ContainsQuery` traits to indicate more specific bounds on registries.
 ### Changed
 - Removed unnecessary generic bounds on `result::Iter`, `result::ParIter`, and `entities::Batch`.